+++ webknossos/webknossos/dataset/dataset.py
@@ -61,7 +61,7 @@ from .sampling_modes import SamplingModes
 if TYPE_CHECKING:
     import pims
 
-    from ..administration.user import Team
+    from ..administration.team import Team
     from ..client._upload_dataset import LayerToLink
     from ..client.context import webknossos_context
 
@@ -3468,7 +3468,7 @@ class RemoteDataset(Dataset):
             - An empty list makes the dataset private
         """
 
-        from ..administration.user import Team
+        from ..administration.team import Team
 
         return tuple(
             Team(id=i.id, name=i.name, organization_id=i.organization)
@@ -3478,7 +3478,7 @@ class RemoteDataset(Dataset):
     @allowed_teams.setter
     def allowed_teams(self, allowed_teams: Sequence[Union[str, "Team"]]) -> None:
         """Assign the teams that are allowed to access the dataset. Specify the teams like this `[Team.get_by_name("Lab_A"), ...]`."""
-        from ..administration.user import Team
+        from ..administration.team import Team
         from ..client.context import _get_api_client
 
         team_ids = [i.id if isinstance(i, Team) else i for i in allowed_teams]